+++ backend/apps/user/models.py
@@ -1,14 +1,5 @@
 from django.db import models
-
-# Create your models here.
-from django.db import models
-from django.conf import settings
-from django.core.mail import send_mail
 from django.contrib.auth.models import BaseUserManager, AbstractBaseUser, PermissionsMixin
-from django.utils import timezone
-from random import randint
-
-from model_utils.models import TimeStampedModel
 
 import datetime
 import uuid
