+++ backend/common/mixins.py
@@ -7,7 +7,7 @@ from rest_framework import status
 from rest_framework.response import Response
 from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
 
-from user.models import User
+from apps.user.models import User
 
 
 class CheckMixin:
