+++ backend/config/urls.py
@@ -18,7 +18,7 @@ schema_view = get_schema_view(
 
 urlpatterns = [
     path('admin', admin.site.urls),
-    path('api/v1/', include('user.urls')),
+    path('api/v1/', include('apps.user.urls')),
 
     # re_path('.*', TemplateView.as_view(template_name='index.tsx')),
 ]
